+++ packages/create-tutorial/CHANGELOG.md
@@ -1,3 +1,11 @@
+## [0.0.3](https://github.com/stackblitz/tutorialkit/compare/1.1.0...0.0.3) "create-tutorial" (2024-10-20)
+
+
+
+## [0.0.3](https://github.com/stackblitz/tutorialkit/compare/1.0.0...0.0.3) "create-tutorial" (2024-10-18)
+
+
+
 ## [0.0.3](https://github.com/stackblitz/tutorialkit/compare/0.2.3...0.0.3) "create-tutorial" (2024-10-01)